satellite-worker/requirements.txt
flask
rasterio
numpy
boto3
pystac-client
shapely
Pillow
pyproj

